+++ pkg/eventStructure/struct.go
@@ -1,0 +1,8 @@
+package eventStructure
+
+
+type Event struct {
+  Count int64
+  Type string
+  Timestamp int64
+}
