+++ src/buffers/answ.proto
@@ -1,0 +1,8 @@
+syntax = "proto3";
+
+package buffers;
+
+message AnswerStruct{
+  int64 Count = 1;
+  string Type = 2;
+}
