+++ src/buffers/struct.proto
@@ -1,0 +1,9 @@
+syntax = "proto3";
+
+package buffers;
+
+message Event {
+  int64 Count = 1;
+  string Type = 2;
+  int64 Timestamp = 3;
+}
